requirements.txt
pandas==1.5.3
Pillow==9.4.0
PyYAML==6.0
torch==1.13.1
torchvision==0.14.1
